workflows/emailWorkflow/requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
beautifulsoup4==4.12.3
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
Deprecated==1.2.16
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
filelock==3.13.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.2.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
guidance==0.1.16
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.3
jiter==0.8.2
joblib==1.4.2
llama-cloud==0.1.10
llama-hub==0.0.79.post1
llama-index==0.12.14
llama-index-agent-openai==0.4.2
llama-index-cli==0.4.0
llama-index-core==0.12.14
llama-index-embeddings-huggingface==0.5.1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.4
llama-index-legacy==0.9.48.post4
llama-index-llms-deepseek==0.1.0
llama-index-llms-huggingface==0.4.2
llama-index-llms-llama-cpp==0.3.0
llama-index-llms-ollama==0.5.0
llama-index-llms-openai==0.3.14
llama-index-llms-openai-like==0.3.3
llama-index-multi-modal-llms-openai==0.4.2
llama-index-program-guidance==0.3.0
llama-index-program-lmformatenforcer==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.4
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.20
llama_cpp_python==0.2.90
llamaindex-py-client==0.1.19
MarkupSafe==2.1.5
marshmallow==3.26.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
ollama==0.4.7
openai==1.60.1
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
pyaml==23.12.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
pypdf==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
text-generation==0.7.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.0+cu124
torchaudio==2.5.0+cu124
torchvision==0.20.0+cu124
tqdm==4.67.1
transformers==4.48.1
triton==3.1.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wrapt==1.17.2
yarl==1.18.3
